fpga-diff: Support the verification function of FPGA difftest
Integrate a project that utilizes FPGA - DIFFTEST. Compress the difftest data and transmit it to the host for processing and comparison. By doing so, the checking efficacy equivalent to that of the emu can be attained.

Performance data:
On our reference machine, the average running speed is 7 MHz. This is dependent on the specific workload and the configuration of the nemu. Generally, the performance fluctuates between 5 and 12 MHz. Higher single-core performance should be able to enhance the performance.

Reference machine: 11th Gen Intel(R) Core(TM) i7-11700 @ 2.50GHz DDR4-3200Mhz

The method of generating the project is the same as that of XSKMH, however, it should be noted that the CPU configuration used should have the "FpgaDiff" prefix.

The usage method of the host computer is as follows: https://docs.xiangshan.cc/zh-cn/latest/tools/difftest/

Co-authored-by: klin02 youkunlin20@mails.ucas.ac.cn